requirements_production.txt
# MLflow Trading System Requirements - Production Ready
# Core ML Libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0

# MLflow - Experiment Tracking
mlflow>=2.0.0

# Data Visualization & Dashboard
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.15.0
streamlit>=1.25.0

# Real-time Data Sources
yfinance>=0.1.87
MetaTrader5>=5.0.37
websocket-client>=1.4.0
requests>=2.28.0

# Deep Learning (Optional)
tensorflow>=2.10.0
torch>=1.12.0

# Web APIs & Services
fastapi>=0.95.0
uvicorn>=0.20.0

# Task Scheduling
schedule>=1.2.0
APScheduler>=3.10.0

# Configuration & Environment
python-dotenv>=0.19.0
pyyaml>=6.0

# Data Processing & Analysis
scipy>=1.11.0
statsmodels>=0.14.0
ta>=0.10.2

# Utilities & Performance
joblib>=1.2.0
tqdm>=4.64.0
psutil>=5.9.0
memory-profiler>=0.60.0

# Development Tools
jupyter>=1.0.0
pytest>=7.0.0
black>=22.0.0

# Data Storage (Optional)
sqlalchemy>=1.4.0
pymongo>=4.0.0

# Trading Specific
# TA-Lib>=0.4.26  # Uncomment if you have TA-Lib installed
